+++ packages/core/strapi/lib/commands/__tests__/data-transfer/export.test.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const { expectExit } = require('./test-utils');
+
 describe('Export', () => {
   const defaultFileName = 'defaultFilename';
 
@@ -54,26 +56,17 @@ describe('Export', () => {
     { virtual: true }
   );
 
-  // other spies=
+  // other spies
   jest.spyOn(console, 'log').mockImplementation(() => {});
   jest.spyOn(console, 'warn').mockImplementation(() => {});
   jest.spyOn(console, 'error').mockImplementation(() => {});
 
-  // Now that everything is mocked, import export command
+  // Now that everything is mocked, load the 'export' command
   const exportCommand = require('../../transfer/export');
 
-  const expectExit = async (code, fn) => {
-    const exit = jest.spyOn(process, 'exit').mockImplementation((number) => {
-      throw new Error(`process.exit: ${number}`);
-    });
-    await expect(async () => {
-      await fn();
-    }).rejects.toThrow();
-    expect(exit).toHaveBeenCalledWith(code);
-    exit.mockRestore();
-  };
-
-  beforeEach(() => {});
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
 
   it('uses path provided by user', async () => {
     const filename = 'test';
