+++ packages/core/strapi/lib/commands/__tests__/data-transfer/test-utils.js
@@ -1,0 +1,16 @@
+'use strict';
+
+const expectExit = async (code, fn) => {
+  const exit = jest.spyOn(process, 'exit').mockImplementation((number) => {
+    throw new Error(`process.exit: ${number}`);
+  });
+  await expect(async () => {
+    await fn();
+  }).rejects.toThrow();
+  expect(exit).toHaveBeenCalledWith(code);
+  exit.mockRestore();
+};
+
+module.exports = {
+  expectExit,
+};
